Implement basic shell

+++ include/ncstd/memory.h
@@ -4,5 +4,5 @@
 
 #define NULL ((void*)0)
 
-void* memset(void* dst, uint8_t value, size_t size);
+void* memset(void* dst, int value, size_t size);
 void* memcpy(void* dst, const void* src, size_t size);
